strands-agent-ecs-mcp-validation-error-workaround/requirements.txt
strands-agents>=1.19.0
fastmcp>=2.13.2
boto3>=1.35.0
pydantic>=2.0.0
